rename taxonomies.py to adjustments.py
